refactor: move repository implementation to infrastructure package
