+++ sqlboiler/infrastructure/employee.go
@@ -1,4 +1,4 @@
-package main
+package infrastructure
 
 import (
 	"context"
